requirements.txt
flask
setuptools
wheel
tqdm
twine
python-jose
